+++ changelog.md
@@ -6,6 +6,7 @@
 
 Now properly tracks responses from item groups other than default and tier token.
 
+Responses are now sorted by number of awards.
 
 ## Bugfixes